+++ ExtendedZDO.cs
@@ -242,6 +242,8 @@ sealed class ExtendedZDO : ZDO
 
         public bool GetSacrifiedCryptKey(long playerID, bool defaultValue = default) => _zdo.GetBool($"player{playerID}_SacrifiedCryptKey", defaultValue);
         public void SetSacrifiedCryptKey(long playerID, bool value) => _zdo.Set($"player{playerID}_SacrifiedCryptKey", value);
+        public bool GetSacrifiedWishbone(long playerID, bool defaultValue = default) => _zdo.GetBool($"player{playerID}_SacrifiedWishbone", defaultValue);
+        public void SetSacrifiedWishbone(long playerID, bool value) => _zdo.Set($"player{playerID}_SacrifiedWishbone", value);
     }
 
     sealed class AdditionalData_(PrefabInfo prefabInfo)